refactor: consolidate single and multi invoice print pages
Extract shared PrintableInvoice + InvoiceBillTo components and invoice print utils (tax columns, base component, watermark) so PrintInvoice and PrintInvoices no longer duplicate ~1300 lines of rendering logic. Add an E2E regression spec covering single- and multi-invoice print rendering.
